Made minor tweaks to the frontend. Rearranging the position of the buttons..for the floating bar
